+++ truffle/src/com.oracle.truffle.polyglot/src/com/oracle/truffle/polyglot/PolyglotEngineImpl.java
@@ -385,7 +385,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
             createInstruments(instrumentsOptions, deprecatedDescriptors);
         }
 
-        validateSandbox();
+        validateOptions();
 
         printDeprecatedOptionsWarning(deprecatedDescriptors);
     }
@@ -795,7 +795,7 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
         for (PolyglotInstrument instrument : instrumentsOptions.keySet()) {
             parseAllOptions(instrument.getEngineOptionValues(), instrumentsOptions.get(instrument), newAllowExperimentalOptions, deprecatedDescriptors);
         }
-        validateSandbox();
+        validateOptions();
         printDeprecatedOptionsWarning(deprecatedDescriptors);
 
         RUNTIME.onEnginePatchSuccess(this.runtimeData);
@@ -2662,6 +2662,13 @@ final class PolyglotEngineImpl implements com.oracle.truffle.polyglot.PolyglotIm
         }
         for (String permittedLanguage : permittedLanguages) {
             idToLanguage.get(permittedLanguage).validateSandbox(sandboxPolicy);
+        }
+    }
+
+    private void validateOptions() {
+        validateSandbox();
+        if (engineOptionValues.get(PolyglotEngineOptions.RelaxStaticObjectSafetyChecks) && engineOptionValues.get(PolyglotEngineOptions.ForceStaticObjectSafetyChecks)) {
+            throw PolyglotEngineException.illegalState("Option engine.RelaxStaticObjectSafetyChecks can not be true at the same time as engine.ForceStaticObjectSafetyChecks.");
         }
     }